Python scrap. Beautiful soup.
Scraping code of a crypto-currencies site, with 100 pages of data.
For each crypto, I took the name, the code and download the logo (in a png file).
The page has a "scrap-protection" code that changes the class of an object, making it more complex to get the info at the first chance. Also, the site collapses if the code runs completely at once, so I put a sleep condition on it.